Do not try to remove a connection the slab twice
A single event loop tick may contain multiple events for a token. If the
connection is reset on the first event, subsequent events will still try
to access that token/connection during the event loop tick. The strategy
now is to mark a connection as reset and remove them only when the event
loop tick is finished.

See https://github.com/carllerche/mio/issues/219 for more details.

Fixes #1