Remove dependency on mio's EventLoop object
The mio API is changing and will no longer have event loop. This patch
updates Server to take the role of the event loop. The Server struct no
longer implements Handler and has a new function called `run` that will
act as an event loop.

Support for timers and channels are missing. However, mob was not using
either of those. Support can/will be added in a future patch.