Remove need for connection state and tick per loop
At the end of each event loop tick the server would do an O(n) scan of
all connections to check their state and remove them from the slab if
they were reset. Now, the connection no longer has to track connection
state and the connection is removed when there is an error.